removed main, all in blueprint now
